删除旧版本
