--- Assets/Quadtree Collider Detection/Step Interpretation/5_Singleton/Resources/QuadtreeWithSingletonSetting.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bd36351b61e6fbe499bdbafd88f1076b, type: 3}
-  m_Name: QuadtreeWithSingletonSetting
-  m_EditorClassIdentifier: 
-  top: 1960
-  right: 1080
-  bottom: 0
-  left: 0
-  maxLeafsNumber: 5
-  minSideLength: 10
