--- Assets/Quadtree Collider Detection/Step Interpretation/6_Upwards/QuadtreeDetectorUpwards.cs
@@ -1,42 +1,0 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
-
-namespace MtC.Tools.Quadtree.Example.Step6Upwards
-{
-    [RequireComponent(typeof(QuadtreeColliderUpwards))]
-    public class QuadtreeDetectorUpwards : MonoBehaviour
-    {
-        QuadtreeColliderUpwards _quadTreeCollider;
-
-        List<GameObject> _colliders = new List<GameObject>();
-
-        private void Awake()
-        {
-            _quadTreeCollider = GetComponent<QuadtreeColliderUpwards>();
-        }
-
-        private void OnEnable()
-        {
-            _quadTreeCollider.collisionEvent += OnQuadtreeCollision;
-        }
-
-        private void OnDisable()
-        {
-            _quadTreeCollider.collisionEvent -= OnQuadtreeCollision;
-        }
-
-        void OnQuadtreeCollision(GameObject collisionGameObject)
-        {
-            _colliders.Add(collisionGameObject);
-        }
-
-        private void OnDrawGizmos()
-        {
-            Gizmos.color = Color.yellow;
-            foreach (GameObject collider in _colliders)
-                if (collider)                           //从碰撞发生到绘制Gizmo中间有很短的时间，如果在这期间物体被销毁了，就获取不到Trnanform出bug，因此要先判断
-                    Gizmos.DrawLine(transform.position, collider.transform.position);
-            _colliders.Clear();
-        }
-    }
-}
