--- Assets/Quadtree Collider Detection/Step Interpretation/6_Upwards/QuadtreeSettingUpwards.cs
@@ -1,14 +1,0 @@
-﻿using UnityEngine;
-
-namespace MtC.Tools.Quadtree.Example.Step6Upwards
-{
-    public class QuadtreeSettingUpwards : ScriptableObject
-    {
-        public float startTop = 1960;
-        public float startRight = 1080;
-        public float startBottom = 0;
-        public float startLeft = 0;
-        public int maxLeafsNumber = 5;
-        public float minSideLength = 10;
-    }
-}
