--- Assets/Quadtree Collider Detection/Step Interpretation/6_Upwards/Resources/QuadtreeCanUpwardsSetting.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 90fa37fcc5c4ed247b4325d96d98a8e9, type: 3}
-  m_Name: QuadtreeCanUpwardsSetting
-  m_EditorClassIdentifier: 
-  startTop: 1000
-  startRight: 1000
-  startBottom: -1000
-  startLeft: -1000
-  maxLeafsNumber: 5
-  minSideLength: 10
