--- Assets/Quadtree_old/Example/ColliderMove.cs
@@ -1,11 +1,0 @@
-﻿using UnityEngine;
-
-public class ColliderMove : MonoBehaviour
-{
-    public float speed { get; set; }
-
-    void Update ()
-    {
-        transform.Translate(transform.up * speed * Time.deltaTime);
-	}
-}
