--- Assets/Quadtree_old/Example/ReloadScene.cs
@@ -1,10 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class ReloadScene : MonoBehaviour
-{
-    public void Reload()
-    {
-        SceneManager.LoadScene(SceneManager.GetActiveScene().name);
-    }
-}
