--- Assets/Quadtree_old/Resources/QuadtreeSetting.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 640c23ffdecd5f342a884c52e36ea3c2, type: 3}
-  m_Name: QuadtreeSetting
-  m_EditorClassIdentifier: 
-  startTop: 1960
-  startRight: 1080
-  startBottom: 0
-  startLeft: 0
-  maxLeafsNumber: 5
-  minSideLength: 10
